// src/types/payRoleSetting.ts
export interface PayRoleSettingsState {
  FixAmount: boolean;
  VariableAmount: boolean;
  TotalAmount: boolean;
  Tds: boolean;
  AnyTax: boolean;
  TravelCost: boolean;
  OtherCost: boolean;
  FoodAllowanceCost: boolean;
  MedicleInsuranceCost: boolean;
  PFAmount: boolean;
  GratuityAmount: boolean;
  LaultyAmount: boolean;
  MisleneousAmount: boolean;
}